requirements.txt
confluent-kafka==2.3.0
requests==2.32.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
requests-cache
pandas
bs4